+++ CHANGES.rst
@@ -6,6 +6,10 @@ _____
 - radices : Remove `RadixBase` class, `BasedReal` implementations now declared on their own
 - tables : Implement `models` features in `HTable`, `HTable.from_model` and models attributes
 
+*Project*
+
+- Poetry usage
+
 0.5.2
 _____